requirements.txt
grpcio
grpcio-tools
requests
matplotlib
uvicorn
fastapi
redis